AI_Automation_Agent/requirements.txt
# AI Automation Agent Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
requests==2.31.0
psutil==5.9.6
loguru==0.7.2
python-dotenv==1.0.0
aiofiles==23.2.1

# Additional utilities
pydantic==2.5.0
python-multipart==0.0.6